feat: tokenを各部で引数で受け渡さないように変更
